backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
openai
pydantic
supabase
python-jose[cryptography]
